[Common] Move IconCache here
The code to do this doesn't need anything specific to smuxi, so let's
move it to Common.

This means that we can't use ProxySettings so move the logic about local
vs remote engine into the frontend's CheckIcon(), which is the one who
currently cares about that.

The code has been relicensed to MIT with permission from meebey, the
only other contributor to this code.
