+++ src/Frontend-GNOME/Frontend-GNOME.csproj
@@ -163,7 +163,6 @@
     <Compile Include="Views\Tags\LinkTag.cs" />
     <Compile Include="Views\Tags\PersonTag.cs" />
     <Compile Include="Views\ChatTreeView.cs" />
-    <Compile Include="IconCache.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="smuxi-frontend-gnome.exe.config">
